BWSR-373: Field class to support alias and nested field
